Small clean up
