+++ source/objects/obj_image.c
