+++ source/objects/obj_quad.c
